Use handler to start containers
This prevents the container from being started and subsequently restarted
by the handler on first run. This would happen because the config would
always report that it was changed and invoke the handler but the container
would be started by the docker_container task and hence get started,
and restarted later on. The docker_container task now uses a state of
present and notifies the handler to start it.